2nd course/2nd semester/Programming language Python/Lab_3-4 (API)/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
Django==5.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
mysqlclient==2.2.4